fix mip status update
