+++ integrity-shield-server/resource/request-handler-config.yaml
@@ -7,6 +7,7 @@ data:
   config.yaml: |
     log:
       level: info
+      manifestSigstoreLogLevel: info
       format: json
     requestFilterProfile: 
       skipObjects:
